+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Planner
 
   * Added World class: [#243](https://github.com/personalrobotics/aikido/pull/243)
+  * Added vector field planner [#246](https://github.com/personalrobotics/aikido/pull/246)
 
 * RViz